quiet some warnings to speed build up

+++ common/FAC/FAC.cpp
@@ -208,7 +208,7 @@ void CFACTransmit::FACParam(CVector<_BINARY>* pbiFACData, CParameter& Parameter)
 
 	(*pbiFACData).ResetBitAccess();
 
-	for (int i = 0; i < NUM_FAC_BITS_PER_BLOCK / sizeof(_BINARY) - 1; i++)
+	for (size_t i = 0; i < NUM_FAC_BITS_PER_BLOCK / sizeof(_BINARY) - 1; i++)
 		CRCObject.AddByte((_BYTE) (*pbiFACData).Separate(sizeof(_BINARY)));
 
 	/* Now, pointer in "enqueue"-function is back at the same place,
@@ -391,7 +391,7 @@ bool CFACReceive::FACParam(CVector<_BINARY>* pbiFACData,
 
 	(*pbiFACData).ResetBitAccess();
 
-	for (int i = 0; i < NUM_FAC_BITS_PER_BLOCK / sizeof(_BINARY) - 1; i++)
+	for (size_t i = 0; i < NUM_FAC_BITS_PER_BLOCK / sizeof(_BINARY) - 1; i++)
 		CRCObject.AddByte((_BYTE) (*pbiFACData).Separate(sizeof(_BINARY)));
 
 	if (CRCObject.CheckCRC((*pbiFACData).Separate(8)) == true)
